feat(mac): performance overlay in the receiver video window
Move PerfOverlay/FlowLayout/BarGraph (pure SwiftUI) from the iOS app
into Shared and pin the same HUD to the bottom of the Mac receiver's
video window, behind a "Performance overlay" toggle in the panel —
same showAnalytics default the iPhone uses. The overlay hosting view
carries the blank-cursor rect so the hidden local pointer stays hidden
over the HUD.
